A general purpose WithCommand option.
