+++ .github/workflows/go.yml
@@ -6,6 +6,9 @@ on:
   pull_request:
     types: [opened, synchronize, reopened]
 
+env:
+  FORCE_COLOR: true
+
 jobs:
 
   build:
@@ -14,20 +17,19 @@ jobs:
     strategy:
       matrix:
         go-version:
-          - '1.18'
+          - '1.22'
+          - '1.23'
     steps:
 
       - name: Set up Go ${{ matrix.go-version }}
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v5
         with:
           go-version: ${{ matrix.go-version }}
         id: go
 
       - name: Check out code into the Go module directory
-        uses: actions/checkout@v2
+        uses: actions/checkout@v4
 
       - name: Test
-        run: go run gotest.tools/gotestsum@v1.8.0 --format testname --
+        run: go run gotest.tools/gotestsum@v1.12.0 --format testname --
           -race -count=1 -short ./...
-        env:
-          FORCE_COLOR: true
