+++ .golangci.yaml
@@ -1,0 +1,39 @@
+run:
+  timeout: 5m
+
+linters:
+  disable-all: false
+  presets:
+    - bugs
+    - unused
+    - complexity
+    - format
+    - performance
+    - style
+  enable:
+    - gci
+  disable:
+    - paralleltest
+    - nlreturn
+    - exhaustivestruct
+    - wsl
+    - godox
+    - scopelint
+    - maligned
+    - interfacer
+    - golint
+    - ireturn
+    - varnamelen
+    - exhaustruct
+    - depguard
+
+issues:
+  exclude-rules:
+    - path: _test\.go
+      linters:
+        - wrapcheck
+
+linters-settings:
+  gomoddirectives:
+    # List of allowed `replace` directives. Default is empty.
+    replace-allow-list: []
