+++ go.mod
@@ -1,15 +1,15 @@
 module github.com/wavesoftware/go-commandline
 
-go 1.18
+go 1.22.0
 
 require (
-	github.com/spf13/cobra v1.5.0
+	github.com/spf13/cobra v1.8.1
 	github.com/wavesoftware/go-retcode v1.0.0
-	gotest.tools/v3 v3.3.0
+	gotest.tools/v3 v3.5.1
 )
 
 require (
-	github.com/google/go-cmp v0.5.5 // indirect
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/google/go-cmp v0.6.0 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 )
